requirements.txt
redis
grpcio
grpcio-tools
grpcio-status
googleapis-common-protos
flask
flask_jwt_extended